style: MemoSplitViewController 내부 네이밍 수정 및 파일명 변경
